% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusTCR.R
\name{ClusTCR}
\alias{ClusTCR}
\title{Creates ClusTCR matrix
This function identifies similar CDR3 amino acid sequences based on the same length and V_gene}
\usage{
ClusTCR(my_file, allele = NULL, v_gene = "v_call", cores_selected = 4)
}
\arguments{
\item{my_file}{uploaded file with junction_aa (CD3 sequences), variable gene.}

\item{allele}{The allele, if present as *00 will be removed if the user requires it.}

\item{v_gene}{Variable gene column name}

\item{cores_selected}{Chose # of cores for parallel functions}
}
\value{
X by Y matrix of strucutrally related CDR3 sequences.
}
\description{
Creates ClusTCR matrix
This function identifies similar CDR3 amino acid sequences based on the same length and V_gene
}
