% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusTCR.R
\name{ClusTCR_Large}
\alias{ClusTCR_Large}
\title{Creates ClusTCR matrix
This function identifies similar CDR3 amino acid sequences based on the same length and V_gene}
\usage{
ClusTCR_Large(my_file, allele = NULL, v_gene = "v_call")
}
\arguments{
\item{my_file}{uploaded file with junction_aa (CD3 sequences), variable gene.}

\item{allele}{The allele, if present as *00 will be removed if the user requires it.}

\item{v_gene}{Variable gene column name}
}
\value{
X by Y matrix of structurally related CDR3 sequences.
}
\description{
Creates ClusTCR matrix
This function identifies similar CDR3 amino acid sequences based on the same length and V_gene
}
