% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netplot.R
\name{ggnet2}
\alias{ggnet2}
\title{Copied code from ggnet's ggnet2 function}
\usage{
ggnet2(
  net,
  mode = "fruchtermanreingold",
  layout.par = NULL,
  layout.exp = 0,
  alpha = 1,
  color = "grey75",
  shape = 19,
  size = 9,
  max_size = 9,
  na.rm = NA,
  palette = NULL,
  alpha.palette = NULL,
  alpha.legend = NA,
  color.palette = palette,
  color.legend = NA,
  shape.palette = NULL,
  shape.legend = NA,
  size.palette = NULL,
  size.legend = NA,
  size.zero = FALSE,
  size.cut = FALSE,
  size.min = NA,
  size.max = NA,
  label = FALSE,
  label.alpha = 1,
  label.color = "black",
  label.size = max_size/2,
  label.trim = FALSE,
  node.alpha = alpha,
  node.color = color,
  node.label = label,
  node.shape = shape,
  node.size = size,
  edge.alpha = 1,
  edge.color = "grey50",
  edge.lty = "solid",
  edge.size = 0.25,
  edge.label = NULL,
  edge.label.alpha = 1,
  edge.label.color = label.color,
  edge.label.fill = "white",
  edge.label.size = max_size/2,
  arrow.size = 0,
  arrow.gap = 0,
  arrow.type = "closed",
  legend.size = 9,
  legend.position = "right",
  ...
)
}
\arguments{
\item{net}{net plot from step 2.}

\item{mode}{= "fruchtermanreingold"}

\item{layout.par}{= NULL,}

\item{layout.exp}{= 0}

\item{alpha}{= 1}

\item{color}{= "grey75"}

\item{shape}{= 19}

\item{size}{= 9}

\item{max_size}{= 9}

\item{na.rm}{= NA}

\item{palette}{= NULL}

\item{alpha.palette}{= NULL}

\item{alpha.legend}{= NA}

\item{color.palette}{= palette}

\item{color.legend}{= NA}

\item{shape.palette}{= NULL}

\item{shape.legend}{= NA}

\item{size.palette}{= NULL}

\item{size.legend}{= NA}

\item{size.zero}{= FALSE}

\item{size.cut}{= FALSE}

\item{size.min}{= NA}

\item{size.max}{= NA}

\item{label}{= FALSE}

\item{label.alpha}{= 1}

\item{label.color}{= "black"}

\item{label.size}{= max_size/2}

\item{label.trim}{= FALSE}

\item{node.alpha}{see \code{alpha}}

\item{node.color}{see \code{color}}

\item{node.label}{see \code{label}}

\item{node.shape}{see \code{shape}}

\item{node.size}{see \code{size}}

\item{edge.alpha}{= 1}

\item{edge.color}{the color of the edges, as a color value, a vector of color
values, or as an edge attribute containing color values.
Defaults to \code{"grey50"}.}

\item{edge.lty}{= "solid"}

\item{edge.size}{= 0.25}

\item{edge.label}{= NULL}

\item{edge.label.alpha}{= 1}

\item{edge.label.color}{= label.color}

\item{edge.label.fill}{= "white"}

\item{edge.label.size}{= max_size/2}

\item{arrow.size}{= 0}

\item{arrow.gap}{= 0}

\item{arrow.type}{= "closed"}

\item{legend.size}{= 9}

\item{legend.position}{= "right"}

\item{...}{Other functions in ggplot2}
}
\value{
A ggplot object displaying the network plot.
}
\description{
Copied code from ggnet's ggnet2 function
}
