% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_CDR3.R
\name{motif_plot}
\alias{motif_plot}
\title{Code for plotting the Motif based on a specific CDR3 length and V gene (see \link{netplot} for )}
\usage{
motif_plot(ClusTCR, Clust_column_name = "cluster", Clust_selected = NULL)
}
\arguments{
\item{ClusTCR}{Matrix file produce from \link{ClusTCR_list_to_matrix}}

\item{Clust_column_name}{Name of clustering column from mcl_cluster file e.g. cluster}

\item{Clust_selected}{Select which cluster to label.}

\item{mcl_cluster}{matrix file produce from \link{mcl_to_lable_clusters}}

\item{selected_col}{Color of selected cluster (Default = purple)}

\item{non_selected_col}{Color of selected cluster (Default = grey80)}

\item{selected_text_col}{Color of selected cluster text (Default = black)}

\item{non_selected_text_col}{Color of selected clusters text (Default = grey40)}

\item{selected_text_size}{Text size of selected cluster (Default = 3)}

\item{non_selected_text_size}{Text size of non-selected clusters (Default = 2)}

\item{label}{Name to display on cluster: Name (CDR3_V_gene_Cluster), cluster, CDR3, V_gene, Len (length of CDR3 sequence)}

\item{select_alpha}{Transparency of non-selected clusters}
}
\description{
Code for plotting the Motif based on a specific CDR3 length and V gene (see \link{netplot} for )
}
