% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netplot.R
\name{netplot}
\alias{netplot}
\title{Code for displaying the network.}
\usage{
netplot(
  ClusTCR,
  Clust_selected = 1,
  selected_col = "purple",
  selected_text_col = "black",
  selected_text_size = 3,
  non_selected_text_size = 2,
  Clust_column_name = "cluster",
  label = c("Name", "cluster", "CDR3", "V_gene", "Len"),
  non_selected_col = "grey80",
  non_selected_text_col = "grey40",
  alpha_selected = 1,
  alpha_non_selected = 0.5,
  colour = "color_test"
)
}
\arguments{
\item{Clust_selected}{Select which cluster to label.}

\item{selected_col}{Color of selected cluster (Default = purple)}

\item{selected_text_col}{Color of selected cluster text (Default = black)}

\item{selected_text_size}{Text size of selected cluster (Default = 3)}

\item{non_selected_text_size}{Text size of non-selected clusters (Default = 2)}

\item{Clust_column_name}{Name of clustering column from mcl_cluster file e.g. cluster (Re-numbering the original_cluster), Original_cluster, Clust_size_order (Based on cluster size e.g. number of nodes)}

\item{label}{Name to display on cluster: Name (CDR3_V_gene_Cluster), cluster, CDR3, V_gene, Len (length of CDR3 sequence),CDR3_selected,Name_selected,cluster_selected, (_selected only prints names of the chosen cluster), None}

\item{non_selected_col}{Color of selected cluster (Default = grey80)}

\item{non_selected_text_col}{Color of selected clusters text (Default = grey40)}

\item{alpha_selected}{Transparency of selected cluster (default = 1)}

\item{alpha_non_selected}{Transparency of non-selected clusters (default = 0.5)}

\item{colour}{Colour selected = "color_test" or all = "color_all"}

\item{mcl_cluster}{list object projected from \link{mcl_cluster}}
}
\description{
Code for displaying the network.
}
