% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcl_clusters.R
\name{mcl_cluster}
\alias{mcl_cluster}
\title{Create the files for labelling the linked clusters from ClusTCR_list_to_matrix function}
\usage{
mcl_cluster(my_file, max.iter = 100, inflation = 1, expansion = 1)
}
\arguments{
\item{my_file}{Matrix file produce from \link{ClusTCR_list_to_matrix}}

\item{max.iter}{Number of iterations to find the steady state of MCL.}
}
\description{
Create the files for labelling the linked clusters from ClusTCR_list_to_matrix function
}
