% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcl_clusters.R
\name{mcl_cluster}
\alias{mcl_cluster}
\title{Create the files for labelling the linked clusters from ClusTCR_list_to_matrix function}
\usage{
mcl_cluster(my_file, max.iter = 10, inflation = 1, expansion = 1)
}
\arguments{
\item{my_file}{Matrix file produce from \link{ClusTCR}}

\item{max.iter}{Number of iterations to find the steady state of MCL.}

\item{inflation}{numeric value}

\item{expansion}{numeric value}

\item{j}{part of MCL clustering}
}
\description{
Create the files for labelling the linked clusters from ClusTCR_list_to_matrix function
}
