% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_from_clusterfile.R
\name{Motif_from_cluster_file}
\alias{Motif_from_cluster_file}
\title{Code for plotting the Motif based on a specific CDR3 length and V gene (see \link{netplot} for ).}
\usage{
Motif_from_cluster_file(ClusTCR, Clust_selected = NULL)
}
\arguments{
\item{ClusTCR}{Cluster file produce from [ClusTCR[\link{2}]]}

\item{Clust_selected}{Select which cluster to reivew}
}
\description{
Code for plotting the Motif based on a specific CDR3 length and V gene (see \link{netplot} for ).
}
