% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_plot.R
\name{motif_plot}
\alias{motif_plot}
\title{Code for plotting the Motif based on a specific CDR3 length and V gene (see \link{netplot} for ).}
\usage{
motif_plot(ClusTCR, Clust_column_name = "cluster", Clust_selected = NULL)
}
\arguments{
\item{ClusTCR}{Matrix file produce from \link{mcl_cluster}}

\item{Clust_column_name}{Name of clustering column from mcl_cluster file e.g. cluster}

\item{Clust_selected}{Select which cluster to display. Only one at a time.}
}
\description{
Code for plotting the Motif based on a specific CDR3 length and V gene (see \link{netplot} for ).
}
